Fix build errors and compiler warnings across multiple files
- CPose3D_unittest: fix Eigen jacobiSvd API for newer Eigen (use JacobiSVD constructor)
- CIbeoLuxETH: use float casts matching convertToCartesian() signature
- ReactiveNav3D_demo: fix sign()*remainder() double/float mixing
- RawLogViewer: fix size_t/float/TCoords conversion warnings
- robot-map-gui: fix int-to-float conversions in GL widget code
- simul-landmarks: fix size_t loop variable for matrix indexing
- vector_with_small_size_optimization: suppress GCC false-positive array-bounds warning
- hwdrivers_kinect example: use float literal for setScale()

+++ apps/mrpt_apps_gui/RawLogViewer/main_imports_exports.cpp
@@ -118,7 +118,7 @@ void xRawLogViewerFrame::OnImportCARMEN([[maybe_unused]] wxCommandEvent& event)
   double maxValidLaserRange;
   strMaxLaserRange.ToCDouble(&maxValidLaserRange);
 
-  float thisTimestamp;
+  double thisTimestamp;
 
   std::vector<std::string> sl;
   (void)mrpt::io::loadTextFile(sl, fil);
@@ -180,7 +180,7 @@ void xRawLogViewerFrame::OnImportCARMEN([[maybe_unused]] wxCommandEvent& event)
       goToTheLastToken(str);
       thisTimestamp = atof(str);
       // Insert timestamp from Carmen logfile, adds UNIX 1970-01-01 offset
-      obsScan->timestamp = mrpt::Clock::fromDouble(static_cast<double>(thisTimestamp));
+      obsScan->timestamp = mrpt::Clock::fromDouble(thisTimestamp);
 
       // thisOdoReading
       if (isTheFirstOdo)
